Outsource media type parsing into dedicated parser object
